fix: remove shared process cache
