+++ apps/watch/pages/[part1]/[part2].tsx
@@ -3,7 +3,7 @@ import { GetStaticPaths, GetStaticProps } from 'next'
 import { ReactElement } from 'react'
 import { VideoContentFields } from '../../__generated__/VideoContentFields'
 import { GetVideoContent } from '../../__generated__/GetVideoContent'
-import { createApolloClient } from '../../src/libs/client'
+import { createApolloClient } from '../../src/libs/apolloClient'
 import { VideoContentPage } from '../../src/components/VideoContentPage'
 import { VideoContainerPage } from '../../src/components/VideoContainerPage/VideoContainerPage'
 import { LanguageProvider } from '../../src/libs/languageContext/LanguageContext'
