+++ apps/watch/pages/[part1]/[part2]/[part3].tsx
@@ -3,7 +3,7 @@ import { GetStaticPaths, GetStaticProps } from 'next'
 import { ReactElement } from 'react'
 import { VideoContentFields } from '../../../__generated__/VideoContentFields'
 import { VideoContentPage } from '../../../src/components/VideoContentPage'
-import { createApolloClient } from '../../../src/libs/client'
+import { createApolloClient } from '../../../src/libs/apolloClient'
 import { GetVideoContainerAndVideoContent } from '../../../__generated__/GetVideoContainerAndVideoContent'
 import { LanguageProvider } from '../../../src/libs/languageContext/LanguageContext'
 import { VideoProvider } from '../../../src/libs/videoContext'
