+++ apps/watch/pages/_app.tsx
@@ -12,7 +12,7 @@ import { CacheProvider } from '@emotion/react'
 import type { EmotionCache } from '@emotion/cache'
 import { createEmotionCache } from '@core/shared/ui/createEmotionCache'
 import { firebaseClient } from '../src/libs/firebaseClient'
-import { useApolloClient } from '../src/libs/client'
+import { useApolloClient } from '../src/libs/apolloClient'
 import '../public/fonts/fonts.css'
 import '../public/styles/carousel.css'
 import '../public/styles/video-js.css'
