+++ apps/watch/pages/api/jf/watch.html/[videoId]/[languageId].ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next'
 import { gql } from '@apollo/client'
-import { createApolloClient } from '../../../../../src/libs/client'
+import { createApolloClient } from '../../../../../src/libs/apolloClient'
 import { GetVideo } from '../../../../../__generated__/GetVideo'
 
 export const GET_VIDEO = gql`
