+++ apps/watch/src/libs/apolloClient/index.ts
@@ -1,0 +1,1 @@
+export { useApolloClient, createApolloClient } from './apolloClient'
