--- apps/watch/src/libs/client/cache/index.ts
@@ -1,1 +1,0 @@
-export { cache } from './cache'
