--- apps/watch/src/libs/client/index.ts
@@ -1,5 +1,0 @@
-import client from './client'
-
-export { useApolloClient, createApolloClient } from './client'
-
-export default client
